requirements.txt
asgiref==3.4.1
Django==3.2.8
django-cors-headers==3.10.0
django-debug-toolbar==3.2.4
django-debug-toolbar-request-history==0.1.4
django-extensions==3.1.5
django-simpleui==2021.11.5
djangorestframework==3.12.4
httpagentparser==1.9.1
mysqlclient==2.0.3
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
typing-extensions==3.10.0.2
